bpf: tests: condense LXC no-backend test
Run the IPv4 and IPv6 test from a single file.
